ocaml-base-compiler 4.11*, ocaml-variants 4.11*: do not pass CC=cc on FreeBSD
As done earlier (in #16575) and mentioned in
https://github.com/ocaml/opam-repository/pull/16568#commitcomment-39656969
this is no longer necessary since https://github.com/ocaml/ocaml/pull/9437 -
which was part of 4.11.0.

NB: I'm not a big fan of modifying existing packages, but imho this is sane
and safe in this case, and looking at #16575 it seemed to be fine with
maintainers back then.

I did not touch:
- 32bit
- flambda+musl
- flambda+musl+static

Since I don't understand how they are supposed to work in the first place.
Eventually it'd make sense to restrict them to supported platforms via
(available = [ os = "linux" & os-distribution = ... ])?