FrontC: fixes building cmxs files and installs cmx files
This PR disables the recently released FrontC 3.4.3 (#18559), as it
was installing incorrectly linked cmxs files that were missing code.
In addition it enables the installation of cmx files.

Since no code was changed but Makefiles only, I decided to pack this
is a patch rather then making a new release. The corresponding patch
is merged into FrontC so in the next release there would not be a need
for such patch.