releases FrontC 4.0.0
This release doesn't really introduce any new features but it switches
FrontC to menhir and dune. The switch to menir was required because
the parser has a circular dependency on the lexer and only menhir is
capable of refactoring tokens from the parser. The circular
dependency, on its own, prevented us from upgrading FrontC to dune,
which in turn prevented us from using FrontC in duniverse.

Switching to menhir required some number of tweaks to the grammar,
which might change the behavior of the parser, hence the new
version. We will still support the old version of FrontC (aka stable)
and release bug fixes, but only the 4.x version will receive new
feautures, such as support for the more modern C syntax.

We also added tests and split the package into three subpackages.
The FrontC package is the library with the same contents as
before. And there are also two packages with executables:

- calipso - a program analysis tool that reduces non-structural
  control flow in C programs;

- ctoxml - a program that will dump the parser C AST in XML.

The `printc` executable is no longer installed as its main purpose is
testing and doesn't provide any useful features for the end users, it
just works as `cat`.