Add conf-capnproto.2.
This is the same as conf-capnproto.1, except the Alpine package has
moved from @testing to @edgecommunity.

This will require a line in /etc/apk/repositories like:
@edgecommunity https://dl-cdn.alpinelinux.org/alpine/edge/community

Note that we are preserving conf-capnproto.1 (even though it no longer
builds with the current Alpine package repos) so that users with an old
mirror can continue to use that.

Signed-off-by: Ewan Mellor <ewan@tarides.com>