releases FrontC 4.1.0
Since the 4.0.0 release we striving to turn FrontC into a robust
parser capable of reading modern C programs. Our immediate goal,
that is mostly achieved is to be able to parse standard C and POSIX
header files available on modern linux systems so that we can parsea
simple C89 programs.

This update received many features from c99 to c11 and a few GNU
extensions but it is still mostly c89 (but we can understand a few
modern features that are used in headers). Otherwise, FrontC is a
pretty robust parser that could be used with real-world C programs.