[coq] Fix the compilation of many Coq plugins for OCaml 4.13
Coq imposes a very strict set of default warnings for plugins plus
`-warn-error`. This was meant to be done only for Coq's CI and
development setup, but this setting wrongly leaked into the release
setup of Coq Opam archive.

We fixup things for OCaml 4.13 while a more complete solution is being
worked on.