Do not depend on `ocamlfind` in `base-bytes`
The current status is probably because `ocamlfind` before 1.5.3 did not know
about bytes, but this leads to `ocamlfind` always being installed when
`base-bytes` is installed - even when `ocamlfind` is not used, e.g. in
setups where everything is built using dune and vendored dependencies
like with opam-monorepo.

This PR changes the dependency into a conflict. Another option could be
to just conflict all `ocamlfind` versions prior to 1.5.3 with
`base-bytes`.