Do not allow co-installation with ounit1 with different version
While building
https://opam.ci.ocaml.org/github/ocaml/opam-repository/commit/aa898f09e7ffe95265651b4b2dda4d21d91baa19/variant/opam-2.1,compilers,4.14,ounit.2.1.2,revdeps,bencode.2.0
it turns out that ounit 2.1.2 was installed via one dependency and
ounit2 2.2.6 via another. This fails to link since the symbols exist
twice.

So while ounit1 should depend on ounit2 after the transition,
ounit2 should fail to install if version of ounit1 does not match its
own version (that is, these versions are known to be compatible).