Add ocaml-option-mingw
This is a "legacy" package, given how long there has been temporary
packaging for OCaml 5 with mingw-w64 support. This package provides
ocaml-option-mingw in a mechanism compatible with the existing
ocaml-option- layout, that it is to say it requires the
ocaml-variants.x.yy.x+options package to be used and conflicts with all
the ocaml-options-only- packages.

Users are expected instead to use system-mingw and either their default
architecture or arch-x86_32/arch-x86_64 to select the mingw-w64 port
(which also works for ocaml-base-compiler).