releases lwt-parallel.1.0.0
The lwt-parallel library allows running lwt computations in different
OS processes. In general, Lwt and Unix fork do not play nicely but
this library makes it possible. See the README and implementation for
more details. The library works on Unix systems, such as GNU/Linux and
macOS.

The bumped to 1.0.0 version indicates the stability of the interface,
i.e., we do not plan to break it without chaning the major number. It
also indicates that the library is ready to use and is supported.

The 1.0.0 release includes some new features, such as an explicit
notation for process snapshots (and support for running multiple
snapshots in parallel). It is also now possible to specify custom
serialization protocols for data that is passed between processes.