Add depopt on dune to batteries.3.6.0
This change has already been upstreamed. Batteries can be built with
dune though this is not the command used to build the package when
installing with opam. Adding this depopt will make it possible to vendor
batteries with opam-monorepo as that tool uses a dependency on dune as
an heuristic to determine if a package can be vendored.