Add `build` field to XML-Light 2.4
The current instructions are missing a `build` field (that
`xml-light.2.3` has). It works because the `install` field uses the
`install_ocamlfind` target which depends on `all` and `opt` but it seems
a bit inelegant to build in the `install` step.

This was detected as opam-monorepo classifies packages without `build`
instructions as packages without a source, which is not the case for
`xml-light`. A PR exists
https://github.com/tarides/opam-monorepo/pull/376 but there is the
question whether missing `build` steps is a packaging mistake or not.

In any case, this PR is a proposal to fix the issue and I'm looking
forward to hear the input of opam-repository maintainers on this
question.