bump conf-pkg-config to 3: remove OpenBSD hack to install a port version
See https://github.com/ocaml/opam-repository/pull/20343 and
https://github.com/mirage/mirage/issues/1262 for the background story:

Mirage 3.x used for "cross-compilation" a pkg-config universe (for opam packages
that contained C libraries). Now, on OpenBSD the system pkg-config was not
powerful enough, and we needed the pkgconfig from ports, and pkg-config < 3
installed a symbolic link to have the pkg-config binary linked to
ports/pkgconfig (which install a binary called pkgconf) in PATH.

Since MirageOS 4 is out of the door since ~1 year, we can finally remove this
hack and to enable forensics / archaeology we put upper bounds to
conf-pkg-config on the relevant packages.