New release of Camlp5 (8.01.00), breaking change in API
The original Camlp5 API for installing new grammars doesn't
allow to detect that somebody has installed a new grammar
twice (which is almost always a bug); this gets fixed.

But several packages relied on the old behaviour, and I
need to send out PRs to fix them.  I'm disabling them
here so that I can relaase this, and then can get to
work on the PRs.