Remove constraints on package versions in conflicts, it's unsupported
After a debugging session with @kit-ty-kate as part of fixing
https://github.com/tarides/opam-monorepo/issues/349 and
https://github.com/ocaml-opam/opam-0install-solver/pull/46 we realized
that OPAM doesn't support resolving package variables in `conflicts`
fields (the linter complains about this in `depends` fileds but not in
`conflicts` fields, but the solver never resolves these variables). So
this constraint doesn't do anything in OPAM and it causes issues in the
opam-0install-solver which attempts to resolve them.