Constrain Links.0.9.7
The installation of Links.0.9.7 with OCaml 5.0 silently fails
(c.f. https://github.com/links-lang/links/issues/1176). This patch
puts an upper bound on the required version of the OCaml compiler such
that OPAM will no longer attempt to install Links.0.9.7 with OCaml
5.0.

We have fixed the compatibility issues with OCaml 5.0 in our
development repository. We will make a separate release to have an
OCaml 5 compatiable package of Links available on OPAM.