Prevent ip2locationio from building with dune 3.13
Dune 3.13 introduces an additional check for the validity of the package
names for opam dependencies thus `lwt.unix` is flagged as invalid.
`lwt.unix` does not make much sense as an OPAM package as it is a
findlib name and available in the `lwt` package.

PR with fix upstream submitted: https://github.com/ip2location/ip2location-io-ocaml/pull/2

Example of failing build:
https://opam.ci.ocaml.org/github/ocaml/opam-repository/commit/fa3e075e57439140df4aacda519ec59df15e0b66/variant/compilers,4.14,dune-configurator.3.13.0~alpha1,revdeps,ip2locationio.1.0.0