Remove conf-zstd.0.1
Simplify the maintenance of this package by only having one version to
contend with. The original addition of conf-zstd.1.3.8 was somewhat
flawed - opam does not support depext version contraints, so packages
trying to encode these should use a different name from the conf
package, to ease the maintenance of the depexts list.

Of the major distributions:
- Ubuntu has had 1.3.8+ since Disco (19.04)
- Fedora has had 1.3.8+ since Fedora 28 (RHEL 8 as well, therefore)
- Alpine has had 1.3.8+ since 3.9

Solvers will already have been picking conf-zstd.1.3.8 as the greatest
version, so this commit explicitly marks conf-zstd.0.1 as unavailable,
and removes the now-redundant version constraint on two old zstandard
packages.