Add mingw-w64-shims.0.1.0
Package ensures the sysroot directories for the two mingw-w64 GCC's in
Cygwin get added to PATH (C++ runtime libraries, GCC runtime libraries,
etc.).

For the internal Cygwin installation, this package also compiles a shim
which is used in the switch directory to promote specific Cygwin
binaries to PATH. This provides the equivalent to conf-msvc32 and
conf-msvc64, as it makes GCC (and cygpath, which is required by flexlink
when targetting mingw-w64) available to ocamlopt from outside opam.