Add Windows system- + host-system- packages
Add the system-mingw and system-msvc packages to specify either the
mingw-w64 or MSVC ports when compiling OCaml and the host-system-mingw
and host-system-msvc packages to be used for the dependency graph.

The intention is to complete this for non-Windows systems, but as there
will always be a chance of an unknown system, host-system-other is added
to be used for all the other ports. This package is significant, as
compiler packages must always install a host-system- package (or the
user could attempt to install another incorrect one).