[new release] grenier (0.15)
CHANGES:

Fix compatibility with OCaml 5:
- balmap: missing functions (`add_to_list`, `to_list`, `of_list`) contributed by @kit-ty-kate
- physh: fix compilation with OCaml 5, contributed by @SquidDev

However physh is disabled for now when compiling with OCaml 5. Thorough reviewing is needed to ensure that the current design is safe with the multicore GC (see [let-def/grenier#10](https://github.com/let-def/grenier/pull/10)).

Add a new "Congre" library, a fast congruence closure algorithm with support for backtracking and interpretability of equivalence classes.